+++ src/main/java/net/kuama/wifiMonitor/WifiLiveData.kt
@@ -2,17 +2,19 @@ package net.kuama.wifiMonitor
 
 import androidx.lifecycle.LiveData
 import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.collect
 import net.kuama.wifiMonitor.data.WifiStatus
 
-class WifiLiveData constructor(private val monitor: WifiMonitor) :
-    LiveData<WifiStatus>() {
-
+@OptIn(ExperimentalCoroutinesApi::class)
+class WifiLiveData constructor(private val monitor: WifiMonitor) : LiveData<WifiStatus>() {
     private var startJob: Job? = null
 
     override fun onActive() {
         super.onActive()
         startJob = CoroutineScope(SupervisorJob() + Dispatchers.Default).launch {
-            monitor.observe(::postValue)
+            monitor.monitor().collect { wifiStatus ->
+                postValue(wifiStatus)
+            }
         }
     }
 
@@ -21,6 +23,5 @@ class WifiLiveData constructor(private val monitor: WifiMonitor) :
         if (startJob?.isActive == true) {
             startJob?.cancel()
         }
-        monitor.stop()
     }
 }