+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## Unreleased
+# Version 1.2.1
+- Changed WifiMonitor.start to monitor
+- Removed stop method
+
+# Version 1.2.0
+- Expose flow of `WifiStatus`
+- Kotlin 1.5
+- Added tests
+- Added WifiMonitor Builder
+- Added Permission Checker class
+- Added an alternative way to build a WifiMonitor object
+
+# Version 1.1.1
+> 2021-06-10
+- Expose current `WiFiStatus`
 
 # Version 1.1.1
 > 2021-06-10